Add throw exception when member is null

+++ user-service/user-service/src/main/java/intelligent_taxi/userservice/command/MemberCommandService.java
@@ -1,15 +1,15 @@
 package intelligent_taxi.userservice.command;
 
+import intelligent_taxi.userservice.controller.restResponse.ResponseMessage;
 import intelligent_taxi.userservice.domain.Member;
 import intelligent_taxi.userservice.dto.changeInfo.ChangeEmailRequest;
 import intelligent_taxi.userservice.dto.changeInfo.ChangePasswordRequest;
-import intelligent_taxi.userservice.dto.changeInfo.WithdrawRequest;
 import intelligent_taxi.userservice.dto.signupAndLogin.MemberLoginRequest;
 import intelligent_taxi.userservice.dto.signupAndLogin.MemberSignupRequest;
+import intelligent_taxi.userservice.exception.MemberCustomException;
 import intelligent_taxi.userservice.jwt.JwtTokenProvider;
 import intelligent_taxi.userservice.jwt.TokenInfo;
 import intelligent_taxi.userservice.repository.MemberRepository;
-import intelligent_taxi.userservice.validator.ServiceValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
@@ -23,29 +23,25 @@ import org.springframework.transaction.annotation.Transactional;
 public class MemberCommandService {
 
     private final MemberRepository memberRepository;
-    private final ServiceValidator serviceValidator;
     private final AuthenticationManagerBuilder authenticationManagerBuilder;
     private final JwtTokenProvider jwtTokenProvider;
 
-    public String signupMember(MemberSignupRequest memberSignupRequest) {
-        serviceValidator.validateDuplicateEmail(memberSignupRequest.getEmail());
-
-        Member member = Member.createMember(memberSignupRequest);
+    public String signupMember(MemberSignupRequest requestDto) {
+        Member member = Member.createMember(requestDto);
         return memberRepository.save(member).getUsername();
     }
 
-    public String signupTaxi(MemberSignupRequest memberSignupRequest) {
-        serviceValidator.validateDuplicateEmail(memberSignupRequest.getEmail());
-
-        Member member = Member.createTaxi(memberSignupRequest);
+    public String signupTaxi(MemberSignupRequest requestDto) {
+        Member member = Member.createTaxi(requestDto);
         return memberRepository.save(member).getUsername();
     }
 
-    public TokenInfo login(MemberLoginRequest memberLoginRequest) {
-        String email = memberLoginRequest.getEmail();
-        String password = memberLoginRequest.getPassword();
+    public TokenInfo login(MemberLoginRequest requestDto) {
+        String email = requestDto.getEmail();
+        String password = requestDto.getPassword();
 
-        Member member = memberRepository.findByEmail(email);
+        Member member = memberRepository.findByEmail(email)
+                .orElseThrow(() -> new MemberCustomException(ResponseMessage.MEMBER_IS_NULL));
         String username = member.getUsername();
 
         UsernamePasswordAuthenticationToken authenticationToken =
@@ -58,31 +54,38 @@ public class MemberCommandService {
                 .generateToken(authentication);
     }
 
-    public void updateEmail(ChangeEmailRequest changeEmailRequest, String username) {
-        String newEmail = changeEmailRequest.getEmail();
-        serviceValidator.validateDuplicateEmail(newEmail);
-
-        Member member = memberRepository.findByUsername(username);
-        member.updateEmail(newEmail);
+    public void updateEmail(ChangeEmailRequest requestDto, String username) {
+        Member member = memberRepository.findByUsername(username)
+                .orElseThrow(() -> new MemberCustomException(ResponseMessage.MEMBER_IS_NULL));
+        member.updateEmail(requestDto.getEmail());
     }
 
-    public void updatePassword(ChangePasswordRequest changePasswordRequest, String username) {
-        Member member = serviceValidator.validatePassword(changePasswordRequest.getOldPassword(), username);
-        member.updatePassword(changePasswordRequest.getNewPassword());
+    public void updatePassword(ChangePasswordRequest requestDto, String username) {
+        Member member = memberRepository.findByUsername(username)
+                .orElseThrow(() -> new MemberCustomException(ResponseMessage.MEMBER_IS_NULL));
+        member.updatePassword(requestDto.getNewPassword(), requestDto.getOldPassword());
     }
 
     public void increaseReport(String username) {
-        Member member = memberRepository.findByUsername(username);
+        Member member = memberRepository.findByUsername(username)
+                .orElseThrow(() -> new MemberCustomException(ResponseMessage.MEMBER_IS_NULL));
         member.increaseReport();
     }
 
     public void cancelBlock(String username) {
-        Member member = memberRepository.findByUsername(username);
+        Member member = memberRepository.findByUsername(username)
+                .orElseThrow(() -> new MemberCustomException(ResponseMessage.MEMBER_IS_NULL));
         member.cancelBlock();
     }
 
-    public void withdrawByUsername(WithdrawRequest withdrawRequest, String username) {
-        Member member = serviceValidator.validatePassword(withdrawRequest.getPassword(), username);
+    public void withdrawByUsername(String username) {
+        Member member = memberRepository.findByUsername(username)
+                .orElseThrow(() -> new MemberCustomException(ResponseMessage.MEMBER_IS_NULL));
+
+        if (!username.equals(member.getUsername())) {
+            throw new MemberCustomException(ResponseMessage.USERNAME_NOT_MATCH);
+        }
+
         memberRepository.delete(member);
     }
 }
